move dummys from testing
