+++ src/context/dummy.cxx
@@ -1,4 +1,17 @@
 module;
 export module libfork.context:dummy_context;
 
-namespace lf {}
+import std;
+
+import libfork.core;
+
+namespace lf {
+
+export struct dummy_context {
+  void post(lf::await_handle<dummy_context>);
+  void push(lf::frame_handle<dummy_context>);
+  auto pop() noexcept -> lf::frame_handle<dummy_context>;
+  auto allocator() noexcept -> dummy_allocator &;
+};
+
+} // namespace lf
